requirements.txt
# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Medical Imaging Frameworks
monai>=1.3.0
medigan>=0.2.0

# Medical Image Processing
nibabel>=5.0.0          # NIfTI files
pydicom>=2.4.0          # DICOM files
SimpleITK>=2.3.0        # Image registration and processing
torchio>=0.19.0         # 3D/4D augmentation

# Computer Vision
opencv-python>=4.8.0
Pillow>=10.0.0
scikit-image>=0.21.0

# Scientific Computing
numpy>=1.24.0
scipy>=1.11.0
pandas>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0
tqdm>=4.66.0

# Metrics and Evaluation
scikit-learn>=1.3.0
pytorch-fid>=0.3.0      # Fréchet Inception Distance (opcional)
lpips>=0.1.4            # Perceptual similarity
torchmetrics>=1.4.0     # FID, IS, SSIM, PSNR

# Jupyter and Interactive
jupyter>=1.0.0
ipykernel>=6.25.0
ipywidgets>=8.1.0
gradio>=4.0.0

# Utilities
pyyaml>=6.0
tensorboard>=2.14.0
wandb>=0.15.0           # Experiment tracking (optional)

# Data Download
kaggle>=1.5.16          # For Kaggle datasets
gdown>=4.7.0            # For Google Drive downloads

# Optional: Advanced Models
diffusers>=0.21.0       # Hugging Face Diffusion models
transformers>=4.33.0    # For foundation models
